Store reps as a number and stop handing accounts to strangers
A real training log kept by hand has half reps in roughly a quarter of its sets:
pushups 14 - 14 - 12 - 10.5, pullups 6 - 6.5 - 5.5. That is the rep that got
most of the way up, and it is the difference between a session that improved and
one that did not. Rounding down understates it and rounding up invents a rep
nobody completed, so reps is numeric everywhere and the stepper moves in halves.

The cost is stated rather than hidden: any change now takes twice the taps. It
is bounded by the prefill starting on last session's number, so the common case
stays zero taps. Rounds still move by one, because half a lap of a circuit is
not a thing anybody writes down.

Signing up with an unmatched email now creates nothing instead of a trainer.
This was never a data leak, and it was verified as one before the change: a
stranger is not staff, owns no clients, and every policy returned nothing for
them. What they got was a permanent account on somebody else's project. The app
already had the right state for it.

The cost there is that a new coach can no longer create their own account. A
trainers row carrying their email has to exist first, which is exactly how Clay
and Alejandra were set up, so it is the path already in use.

85 tests pass, up from 82.

+++ js/schema.js
@@ -198,7 +198,12 @@ export const TABLES = {
       // Null on a carry or a sled, where the load is the whole point and there are no reps to
       // count. Anything that computes volume or an estimated 1RM skips these rows rather than
       // inventing a number for them.
-      reps: { type: INT, nullable: true },
+      //
+      // Numeric, not an integer, because a half rep is a real thing people write down: the rep
+      // that got most of the way up. Rounding it away throws out the difference between a
+      // session that improved and one that did not, and rounding it up invents a rep nobody
+      // completed. Epley and volume are both indifferent to it.
+      reps: { type: NUMERIC, nullable: true },
       // Rounds completed in an AMRAP block. Null everywhere else.
       rounds: { type: INT, nullable: true },
       rpe: { type: NUMERIC, nullable: true },
